Update pyramidalization.awk
In the previous version, sin(thetap) was calculated in the function thetap. For the previous version, if an ideally sp2-hybridized carbon atom was present, "-nan" would appear in the output text and a message reading "warning: sqrt: received negative argument -3.28955e-17" would appear in the terminal emulator. In the current version, the square of sin(thetap) is calculated in the function thetap. For the current version, when the absolute value of square of sin(thetap) is smaller than 1e-6, the square of sin(thetap) will be modified to zero.